worldplay-worker/requirements.txt
# WorldPlay Worker Dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# Data validation
pydantic==2.5.3

# Redis
redis==5.0.1

# Image/Video processing
numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Video encoding (optional - for H.264)
av>=11.0.0

# PyTorch (for HY-WorldPlay)
# Install separately based on CUDA version:
# pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu118

# HY-WorldPlay dependencies (from HY-WorldPlay/requirements.txt)
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
einops>=0.7.0
omegaconf>=2.3.0
sentencepiece>=0.1.99
